modularize: rpc handling
